[java] Add extra info about using multiple clients
The AsyncKuduClient and KuduClient are meant to map 1-1 with Kudu
clusters used by an application. In particular, an application shouldn't
make more than one client per cluster. This patch adds a note to this
effect to the AsyncKuduClient javadoc.

Additionally, some use cases require using multiple clients in multiple
applications running on the same machine. For example, to provide very
strict classpath isolation, or to run applications on different JVM
versions, or even for extra security by isolating different data flows
in different processes. In this case, I've seen client instances put
pressure on thread limits on machines with a lot of cores because the
AsyncKuduClient starts at least (2 * #cpu) threads for a netty threadpool.
This patch also adds advice to adjust the netty worker count to help
users with this case.

Change-Id: I34242862ee6d806b74d7717399e5fd0260dece89
Reviewed-on: http://gerrit.cloudera.org:8080/10962
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins