schema: add is_deleted virtual column
This patch introduces a very basic concept of a "virtual column". Virtual
columns borrow from other databases in that they are columns that, rather
than being backed by physical data, are instead backed by Kudu itself. They
may not be part of a schema during table creation/alteration, but may be
added to projections during a scan.

Kudu's virtual columns are defined as logical data types. As data types are
not user-defined, there's no danger of a "collision" between a virtual
column and a physical column as there would be if a virtual column occupied
a well-defined name.

A Kudu subsystem on the scan path that wishes to interact with a virtual
column needs to first figure out if the projection includes it. When
projected, the virtual column's data will be either some default or null
(depending on exactly how it was defined in the projection); it's the
subsystem's responsibility to fill in something meaningful afterwards.

Beyond the basic definition, this patch introduces an IS_DELETED virtual
column derived from BOOL. IS_DELETED will be used to support incremental
backups by describing whether a row was deleted between two timestamps.

Change-Id: I56eb1d44ba8bfbd76d8bb794b8076b695782939e
Reviewed-on: http://gerrit.cloudera.org:8080/10968
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>