memrowset: support iteration with is_deleted virtual column
This patch rounds out the MemRowSet changes for incremental backups.
Taken together, it is now possible to iterate on a specific time range and
to learn which rows were deleted during that time range.

Data type based virtual columns show a wart here: the virtual column definition
itself does not specify whether it is nullable or has a read default value. To
simplify, we enforce that it must have a read default value.

Change-Id: Ic6b053f5a3696eb9d7c26b8e3d96752f4f87bcd8
Reviewed-on: http://gerrit.cloudera.org:8080/10990
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>