[gradle] Add rerunTests property
Gradle keeps track of changes and will skip tasks
that are up-to-date when possible. This is great
for incremental build speed, but can work against you
if you want to rerun tests multiple times.

Gradle does have a —rerun-tasks flag that will make
sure the tests are run. However, it will also rerun all
the tasks in the build tree.

This patch adds a rerunTests property that will
tell gradle the tests need to be rerun when passed.

Change-Id: I88eebdc34c5b9763b87879eade3dc9b42d3caf1c
Reviewed-on: http://gerrit.cloudera.org:8080/11035
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Grant Henke <granthenke@apache.org>