[Java] Retry tests that don’t inherit from BaseKuduTest
The Scala tests and a couple special case tests
don’t inherit from BaseKuduTest but may still
be flaky.

To fix this I moved the retry configuration logic into
the RetryRule and added the rule to the a few tests directly:
   - TestSecurity
   - TestKuduMetastorePlugin

Additionally I updated the Scala TestContext,
now called KuduTestSuite, to
use the JUnitSuite and include the RetryRule so
that all of the Scala tests would be retried. This
required a few syntax changes but has the added
benefit of not needing
`@RunWith(classOf[JUnitRunner])` at the top of
each test class. It also follows similar syntax to the
rest of our tests.

Change-Id: I596fde91a13f7e644fd1b1814e9a672f96fa0e4b
Reviewed-on: http://gerrit.cloudera.org:8080/11037
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins