deltas: fuzz tests for delta file and dms
Here are new fuzz tests for deltafile-test and deltamemstore-test that
generate random deltas and iterate over them using random snapshots, testing
the various delta iterator functions. I found this to be a simpler
environment in which to debug delta-related issues, vs. more heavy-weight
tests (i.e. fuzz-itest).

The test introduces MirroredDeltas, a utility that's useful for tracking
deltas that are written into a delta file or DMS.

Change-Id: I539ff0f2055cf398a42efb824238188230e25516
Reviewed-on: http://gerrit.cloudera.org:8080/11140
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>