[catalog_manager] re-replication warning message update
Updated the warning message upon creating a table with replication
factor F when only (F - 1) tablet servers are alive.

Do not mention unsetting --raft_prepare_replacement_before_eviction
since the 3-4-3 replica management scheme allows to spawn a new
replica in-place in case of unrecoverable failure such as IO error
and falling behind WAL's ancient history mark.  So, the 3-2-3 scheme
is no better in that regard and would not be able to re-replicate
in case of a tablet server failure either.  Also, the 3-2-3 scheme
is deprecated at this point.

I also did few other petty changes around the warning message.

There are no functional changes in this patch.

Change-Id: I608ce179548ac25d660d713e7ce1d0e067dd4e02
Reviewed-on: http://gerrit.cloudera.org:8080/11114
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>