build: allow overriding of pip installation parameters
This commit introduces the PIP_INSTALL_FLAGS environment variable to
build-and-test.sh which can be used to override various pip installation
behaviors. For example, you can pass "-i <URL>" so that pip uses a different
installation URL to retrieve packages.

For this to work effectively, PIP_INSTALL_FLAGS must come after all of
build-and-test.sh's pip installation flags. That's because if the same flag
is passed multiple times on the command line, the last value "wins". To make
this consistent, I modified MVN_FLAGS and GRADLE_FLAGS to behave the same.

I also copied the pytest-timeout dependency from setup.py to
requirements.txt so that "python setup.py test" won't have to download
anything if it was preceeded with "pip install -r requirements.txt". It is
possible to configure setuptools with a custom index URL instead, but the
various[1] approaches[2] involve modifying files rather than passing
command line arguments.

1. https://setuptools.readthedocs.io/en/latest/easy_install.html#configuration-files
2. https://setuptools.readthedocs.io/en/latest/setuptools.html (search for dependency_links)

Change-Id: Ib09da4bb2e2cdcbf7ea4a8ab2c1089ab46d934fb
Reviewed-on: http://gerrit.cloudera.org:8080/11116
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>