[consensus] KUDU-2335 increment term on explicit step down
Prior to this patch, the catalog manager could get a tablet report
from a former leader replica with empty leader UUID and old term.
A dedicated logic in the catalog manager's code (see section 7d(i))
would amend the empty leader UUID to replace it with the previous
leader's UUID.  As a result of those shenanigans, catalog manager
would interpret the incoming report as a report from a leader replica
that reports its own health status as UNKNOWN.

The TwoConcurrentRebalancers scenario of the recently introduced
ConcurrentRebalancersTest reproduces the issue pretty often
(about 1 in 100 runs failed), so it was easy to pin-point the problem.
Mike sketched the fix and I ran the new code via dist-test about 1K
times and verified the problem is gone.

As for the test coverage, in addition to the already mentioned
would-be-flaky TwoConcurrentRebalancers scenario, I modified
RaftConsensusElectionITest.LeaderStepDown to reliably catch regressions.

Additionally, I updated the TabletCopyITest.TestRejectRogueLeader
scenario not asking rogue leader to step down: otherwise it would
fail.

Change-Id: I4e1f1446176a78ba04e74dd1153f9048a32d8d5f
Reviewed-on: http://gerrit.cloudera.org:8080/11122
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Alexey Serbin <aserbin@cloudera.com>