Add unstable client APIs to fetch HMS integration configuration status
The Impala/Kudu integration must know whether the Kudu cluster is
configured with the HMS integration enabled. This commit adds an
unstable API to the Java client and a private API to the Kudu client to
lookup this information. The new C++ API is utilized by the HMS tools to
get the HMS connection information without having to use the GetFlags
API, which requires admin privileges.

The patch required adding support for enabling the HMS integration to
the Java mini cluster.

Change-Id: Iddb89787ed35c41e85f1d9bf953c4c228dcafcdb
Reviewed-on: http://gerrit.cloudera.org:8080/11121
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins