[gradle] Remove integrationTest tasks
When initially adding the Gradle build I added separate
source set definitions and tasks for Integration tests to
match the Maven build. However, given that the
gradle retry logic is the same for all tests and there
is no real difference between and integration test and
a unit test, we should remove this to simplify the build.

Change-Id: I58c0bafd19cc21ad9c0f83159fdc47d01e3731bf
Reviewed-on: http://gerrit.cloudera.org:8080/11130
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins