KUDU-2411. Add scripts to build binaries for testing use
This script will build a jar file containing "relocatable" binaries for
use with a minicluster. The idea is that the resulting jar will be
uploaded to Maven Central whenever there is a release. There should be
one artifact built on Linux (preferably an old OS, such as RHEL 6) and
one on macOS, so that people can run integration tests against Kudu on
either platform.

The current approach is to dynamically link the dependencies instead of
statically linking them to optimize for the total size of the resulting
test artifact. With kudu-tserver, kudu-master, and kudu admin tool
binaries all linking against the same shared objects, the size of the
jar file to download and unpack is significantly reduced. There are a
bunch of shared objects we have to ship regardless, such as dependencies
from thirdparty and security libraries from the system.

Other pieces not included in this patch are code to depend on, locate,
and extract the resulting JAR file, as well as a refactor of the
MiniCluster code to use the resulting bits to start up a Kudu cluster as
part of a Java integration test. Some of that work is underway or has
already been merged as part of KUDU-2411.

Very basic testing of this patch has been done on both Linux and macOS.
We will need to do more testing and validation as the integration work
progresses.

Change-Id: I8b8f90cfe80f6830177bf6ea9e0711eb0d034f75
Reviewed-on: http://gerrit.cloudera.org:8080/11377
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins