[build] Switch to sha512 checksum for releases
According to the current release distribution guide[1] the checksum file
SHOULD be SHA-256 and/or SHA-512 and SHOULD NOT be SHA-1

SHA-512 seemed more future-proof than SHA-256

Another change is the separator between the checksum and the filename.
Previously it was a <TAB> character, but at least on MacOS the tab
version doesn't work when trying to verify the checksum with `shasum -c
<checksum_file>`

This commit changes it to double-space which works on both Linux using
`sha512sum` and MacOS using `shasum`.

Looking at a checksum created by `sha512sum` on Linux, it also uses
double-space as the separator, so this seems to be the de-facto
standard:

$ sha512sum apache-kudu-1.8.0-SNAPSHOT.tar.gz | xxd
0000070: 3065 3132 3666 3463 3664 6332 3065 6338  0e126f4c6dc20ec8
0000080: 2020 6170 6163 6865 2d6b 7564 752d 312e    apache-kudu-1.

(0x20 is <SPACE>[2])

[1] http://www.apache.org/dev/release-distribution#sigs-and-sums
[2] man 7 ascii

Change-Id: I7646b4559bf39d2415b32a1dcdd4cd7ecde41531
Reviewed-on: http://gerrit.cloudera.org:8080/11217
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Todd Lipcon <todd@apache.org>