Tag follower_unavailable_considered_failed_sec as runtime safe
This patch tags the
follower_unavailable_considered_failed_sec flag as
runtime safe so that the set_flags tools do not
require the -force argument.

This flag has been defacto runtime changable since
at least Kudu 1.5.0 but hasn’t been marked as such.

It is runtime changeable because the flag is checked
each time the peer health status is checked.

Change-Id: I3328f7824a6abfff017bc05ae5b053a9de9a834f
Reviewed-on: http://gerrit.cloudera.org:8080/11261
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>