[webui] Make the tables on /tablets fancy
The /tablets page has live tablets and a tombstoned tablets tables.
These tables get big and unwieldy to use. This patch makes the tables
paginated and searchable by adding the bootstrap-table plugin[1]. I
picked this plugin because it integrates with our existing bootstrap
styles, it worked out of the box with no effort, and the project looks
active. I would've liked a plugin that could be used to sort columns
based on their type out of the box, but I couldn't find a good
candidate.

I also removed a couple of columns that I've never seen be useful:

* Write buffer memory usage: Users are not usually interested in this for
  specific tablets, except possibly to identify outliers, which can be
  done on the mem-trackers page.
* Last status: This is updated at tablet creation or bootstrap time and
  never contains useful information.

Less columns allows the table to fit comfortably on smaller screens.

1: https://github.com/wenzhixin/bootstrap-table

Change-Id: I5243e9cb839e75936c71b14fed08881317ee0f5f
Reviewed-on: http://gerrit.cloudera.org:8080/11290
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>