util: add once class based on std::call_once
The existing KuduOnceDynamic class only allows calling functions of type
Status(void). I intend on adding an argument to some of the existing the
Status(void) methods that get called using a KuduOnceDynamic. As such,
this patch adds a KuduOnceLambda class that accepts a lambda and
otherwise has the same semantics as KuduOnceDynamic.

I've replaced the KuduOnceDynamics that call methods that I intend on
addding arguments to with their corresponding KuduOnceLambdas.

Change-Id: Ide56057a800bf07923031df5a2a76a42f0b15358
Reviewed-on: http://gerrit.cloudera.org:8080/11302
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>