KUDU-2012 Kudu Flume sink auth support
Adding FlumeAuthenticator to KuduSink and creating KuduClient
inside a PrivilegedExecutor action.

Added an extra step to the mini cluster to create
a keyTab for the client used for testing.

Added automated test with short KDC ticket lifetime
to test reacquiring.

Manual testing was done on a secure cluster as well.

Change-Id: I11b5f08802883afa178d346af48d3bcd15281917
Reviewed-on: http://gerrit.cloudera.org:8080/11334
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>