cfile-test: pass in IOContext on Open()
Previously, on certain machines, cfile-test's TestDataCorruption would
fail 100% of the time when attempting to read from a null IOContext. The
passed-in ReaderOptions struct was constructed with the default
constructor, and because no default was set for its io_context member,
this resulted in undefined behavior, which explains why this wasn't
caught in pre-commit.

This patch updates cfile-test to correctly pass in an IOContext, and
assigns a default for ReaderOption's io_context.

Change-Id: Iaed883d26075ed1823882a7f7b18834f39b0732d
Reviewed-on: http://gerrit.cloudera.org:8080/11389
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins