KUDU-686 (part 1/2): decompose guts of DMSIterator into DeltaPreparer
To address KUDU-686, we're going to repurpose DMSIterator's PrepareBatch()
machinery and associated in-memory state for use in the DeltaFileIterator.
Doing so obviates the need for a "multi-pass" approach to ApplyUpdates()
because DeltaFileIterator will no longer be performing any decoding there,
having done all of it in PrepareBatch().

This patch lays the groundwork by refactoring the guts of DMSIterator into
the new DeltaPreparer class. DMSIterator will continue to concern itself
with CBTree iteration, but will delegate the delta preparation and service
to DeltaPreparer.

In performing this refactor, I tried to be as faithful as possible to the
original code. The one exception is that I replaced prepared_idx_ and
prepared_count_ with state that I found easier to understand.

No new tests; I figured there was enough test coverage of DMSIterator, and
testing DeltaPreparer directly seemed like it'd be low bang for the buck.

Change-Id: I26c92dfa69b5323e7fbb18d02010ed99cc29c3e5
Reviewed-on: http://gerrit.cloudera.org:8080/11394
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>