[tools] Fix help text for replica_type
'kudu tablet change_config [add_replica|change_replica_type]' take an
argument for 'replica type', which the help text said had to be "VOTER"
or "NON-VOTER", but actually it has to be "VOTER" or "NON_VOTER",
because it is parsed into an enum by a generated PB function, and the PB
enum names are "VOTER" and "NON_VOTER". This fixes the help text.

Change-Id: Icb43e0eaffbfa17967103c7aebe15ea0a06c5598
Reviewed-on: http://gerrit.cloudera.org:8080/11400
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>