[tools] ksck checksums: Factor out of main ksck code
To prepare to address KUDU-2179 and to improve the organization of the
code, this patch factors checksum-related classes and code out of
ksck.{cc,h} into a new pair of files ksck_checksum.{cc,h}. Following
the pattern of other ksck-related classes, it also renames some classes
so they start with "Ksck".

There are no functional changes in this patch.

Change-Id: I4bb1f51af22ab0c6c20b9426dbb62ea48413ed5b
Reviewed-on: http://gerrit.cloudera.org:8080/11488
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins