[jepsen] a bunch of minor updates on README.adoc
Added a link to the reference script for building Kudu and running
Jepsen tests.  Added information on how to quickly locate corresponding
sections in the Jepsen output in case if an error happened during Jepsen
tests run.  Updated the command to build the kudu-jepnsen package.

This patch does not contain any functional changes.

Change-Id: Ibd4c7a522829e774f786a85d4bdfa01d0be553f5
Reviewed-on: http://gerrit.cloudera.org:8080/11414
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>