(02/05) delta_store: convert DeltaIterator::PrepareBatch flags into bitfield
A follow-on patch will introduce DeltaIterator::SelectUpdates(), a new
method for extracting delta information from a DeltaIterator. As it is
only to be used by diff scans and will require dedicated in-memory state,
it'll also come with a new DeltaIterator::PrepareBatch() flag.

However, diff scans will need to use both SelectUpdates() and
ApplyUpdates(), and it'd be a shame to have to reseek and reprepare just to
use both extraction methods. As such, this patch makes it possible to
prepare a batch for use in multiple delta extraction methods.

Change-Id: Id500ec3cc9459a78ae1098be3f3cd0cacb7a7a1a
Reviewed-on: http://gerrit.cloudera.org:8080/11857
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>